[mount-sd] create directory writable for all users. Fixes MER#1010
If card has filesystem different from vfat or ntfs, so maybe posix attrs are
applicable to it, create directory there with recognizable name and sticky bit
to allow users (e.g. nemo) to create files/directories there. To git rid of the
hassle of handling card with different filesystems in a different way and
hard-coding sticky-marked dir name, symlink this directory for posix-compliant
filesystems in the temporary runtime dir /run/media or just the filesystem root
directory - for vfat and ntfs

Signed-off-by: Denis Zalevskiy <denis.zalevskiy@jolla.com>